# Stack-based unmarshaler