Instead of tracking the build chain using `target`s, use the stack.

Working and most benchmarks look good, ~but regression on structs unmarshalling (~50% slower on `ReferenceFile/struct`)~ (was actually a bug in v2, see https://github.com/pelletier/go-toml/pull/547).

---

### Comparison against v1 and BurntSushi/toml

```
./ci.sh benchmark -a
```

```
name \ time/op                   go-toml-v2.txt  go-toml-v1.txt  bs-toml.txt
UnmarshalDataset/config-8           26.0ms ± 1%    113.9ms ± 1%    73.7ms ±15%
UnmarshalDataset/citm_catalog-8     30.7ms ± 3%     61.0ms ± 2%    40.1ms ± 1%
UnmarshalDataset/twitter-8          12.7ms ± 1%     31.9ms ± 1%    21.4ms ± 1%
UnmarshalDataset/code-8              101ms ± 1%      238ms ± 3%     241ms ± 2%
UnmarshalDataset/example-8           214µs ± 1%      633µs ± 1%     441µs ± 1%
UnmarshalSimple/struct-8             756ns ± 0%     3925ns ± 2%    2452ns ± 1%
UnmarshalSimple/map-8               1.06µs ± 1%     3.97µs ± 2%    2.67µs ± 1%
ReferenceFile/struct-8              47.8µs ± 3%    270.5µs ± 2%   311.2µs ± 1%
ReferenceFile/map-8                 74.7µs ± 1%    241.3µs ± 1%   244.0µs ± 4%
HugoFrontMatter-8                   16.1µs ± 0%     45.3µs ± 0%    37.4µs ± 1%
[Geo mean]                           316µs          1044µs          812µs   
```


---


### Comparison against v2

```
go test -run=Nothing ./benchmark -bench=. -count=10
```

```
name                              old time/op    new time/op    delta
UnmarshalDataset/config-32          38.2ms ± 0%    20.2ms ± 0%  -47.07%  (p=0.000 n=10+10)
UnmarshalDataset/canada-32          72.7ms ± 0%    65.6ms ± 1%   -9.84%  (p=0.000 n=9+10)
UnmarshalDataset/citm_catalog-32    31.8ms ± 1%    22.1ms ± 1%  -30.48%  (p=0.000 n=10+10)
UnmarshalDataset/twitter-32         13.8ms ± 0%     8.7ms ± 1%  -37.39%  (p=0.000 n=10+10)
UnmarshalDataset/code-32             161ms ± 1%      84ms ± 1%  -47.97%  (p=0.000 n=10+10)
UnmarshalDataset/example-32          277µs ± 0%     158µs ± 0%  -42.95%  (p=0.000 n=10+10)
UnmarshalSimple/struct-32            605ns ± 0%     561ns ± 0%   -7.18%  (p=0.000 n=10+10)
UnmarshalSimple/map-32              1.05µs ± 0%    0.80µs ± 0%  -24.34%  (p=0.000 n=10+10)
ReferenceFile/struct-32             48.4µs ± 0%    35.2µs ± 0%  -27.39%  (p=0.000 n=10+10)
ReferenceFile/map-32                 104µs ± 0%      56µs ± 0%  -45.89%  (p=0.000 n=7+10)
HugoFrontMatter-32                  20.5µs ± 0%    12.6µs ± 0%  -38.41%  (p=0.000 n=10+10)

name                              old speed      new speed      delta
UnmarshalDataset/config-32        27.4MB/s ± 0%  51.8MB/s ± 0%  +88.93%  (p=0.000 n=10+10)
UnmarshalDataset/canada-32        30.3MB/s ± 0%  33.6MB/s ± 1%  +10.91%  (p=0.000 n=9+10)
UnmarshalDataset/citm_catalog-32  17.5MB/s ± 1%  25.2MB/s ± 1%  +43.85%  (p=0.000 n=10+10)
UnmarshalDataset/twitter-32       31.9MB/s ± 0%  51.0MB/s ± 1%  +59.73%  (p=0.000 n=10+10)
UnmarshalDataset/code-32          16.6MB/s ± 1%  32.0MB/s ± 1%  +92.21%  (p=0.000 n=10+10)
UnmarshalDataset/example-32       29.2MB/s ± 0%  51.2MB/s ± 0%  +75.28%  (p=0.000 n=10+10)
UnmarshalSimple/struct-32         18.2MB/s ± 0%  19.6MB/s ± 0%   +7.73%  (p=0.000 n=10+10)
UnmarshalSimple/map-32            10.4MB/s ± 0%  13.8MB/s ± 0%  +32.15%  (p=0.000 n=10+8)
ReferenceFile/struct-32            108MB/s ± 0%   149MB/s ± 0%  +37.72%  (p=0.000 n=10+10)
ReferenceFile/map-32              50.3MB/s ± 0%  93.0MB/s ± 0%  +84.79%  (p=0.000 n=7+10)
HugoFrontMatter-32                26.6MB/s ± 0%  43.2MB/s ± 0%  +62.37%  (p=0.000 n=10+10)

name                              old alloc/op   new alloc/op   delta
UnmarshalDataset/config-32          13.9MB ± 0%     7.6MB ± 0%  -45.36%  (p=0.000 n=9+10)
UnmarshalDataset/canada-32          74.3MB ± 0%    84.4MB ± 0%  +13.59%  (p=0.000 n=9+9)
UnmarshalDataset/citm_catalog-32    37.3MB ± 0%    35.7MB ± 0%   -4.46%  (p=0.000 n=10+10)
UnmarshalDataset/twitter-32         15.6MB ± 0%    14.0MB ± 0%  -10.55%  (p=0.000 n=10+10)
UnmarshalDataset/code-32            52.4MB ± 0%    23.2MB ± 0%  -55.76%  (p=0.000 n=10+9)
UnmarshalDataset/example-32          239kB ± 0%     202kB ± 0%  -15.51%  (p=0.000 n=10+10)
UnmarshalSimple/struct-32             709B ± 0%      661B ± 0%   -6.77%  (p=0.000 n=10+10)
UnmarshalSimple/map-32              1.17kB ± 0%    1.04kB ± 0%  -11.59%  (p=0.000 n=10+10)
ReferenceFile/struct-32             24.8kB ± 0%    13.0kB ± 0%  -47.58%  (p=0.000 n=10+10)
ReferenceFile/map-32                53.5kB ± 0%    30.4kB ± 0%  -43.10%  (p=0.000 n=10+10)
HugoFrontMatter-32                  11.1kB ± 0%     8.6kB ± 0%  -22.71%  (p=0.000 n=10+10)

name                              old allocs/op  new allocs/op  delta
UnmarshalDataset/config-32            557k ± 0%      259k ± 0%  -53.54%  (p=0.000 n=9+10)
UnmarshalDataset/canada-32            896k ± 0%      782k ± 0%  -12.74%  (p=0.000 n=9+9)
UnmarshalDataset/citm_catalog-32      377k ± 0%      193k ± 0%  -48.82%  (p=0.000 n=10+9)
UnmarshalDataset/twitter-32           158k ± 0%       62k ± 0%  -60.82%  (p=0.000 n=10+9)
UnmarshalDataset/code-32             2.57M ± 0%     1.08M ± 0%  -58.12%  (p=0.000 n=10+9)
UnmarshalDataset/example-32          3.64k ± 0%     1.48k ± 0%  -59.40%  (p=0.000 n=10+10)
UnmarshalSimple/struct-32             10.0 ± 0%       8.0 ± 0%  -20.00%  (p=0.000 n=10+10)
UnmarshalSimple/map-32                19.0 ± 0%      13.0 ± 0%  -31.58%  (p=0.000 n=10+10)
ReferenceFile/struct-32                461 ± 0%       204 ± 0%  -55.75%  (p=0.000 n=10+10)
ReferenceFile/map-32                 1.49k ± 0%     0.67k ± 0%  -55.00%  (p=0.000 n=10+10)
HugoFrontMatter-32                     313 ± 0%       160 ± 0%  -48.88%  (p=0.000 n=10+10)
```