@pelletier 2021-05-30
Noticed that `ast.(*Node).Key()` has almost doubled in cum% cpu time. There was no change in that area, so as expected the flat cost is the sam (which is relatively high). It suggests it is called about twice as much.

Verifying with:

```diff
diff --git a/benchmark/benchmark_test.go b/benchmark/benchmark_test.go
index 1337895..a252781 100644
--- a/benchmark/benchmark_test.go
+++ b/benchmark/benchmark_test.go
@@ -6,6 +6,7 @@ import (
 	"time"
 
 	"github.com/pelletier/go-toml/v2"
+	"github.com/pelletier/go-toml/v2/internal/ast"
 	"github.com/stretchr/testify/require"
 )
 
@@ -160,6 +161,7 @@ func BenchmarkReferenceFile(b *testing.B) {
 		b.SetBytes(int64(len(bytes)))
 		b.ReportAllocs()
 		b.ResetTimer()
+		ast.Count = 0
 		for i := 0; i < b.N; i++ {
 			d := benchmarkDoc{}
 			err := toml.Unmarshal(bytes, &d)
@@ -167,6 +169,7 @@ func BenchmarkReferenceFile(b *testing.B) {
 				panic(err)
 			}
 		}
+		b.ReportMetric(float64(ast.Count)/float64(b.N), "keyCall/op")
 	})
 
 	b.Run("map", func(b *testing.B) {
diff --git a/internal/ast/ast.go b/internal/ast/ast.go
index ba2729e..e69bef2 100644
--- a/internal/ast/ast.go
+++ b/internal/ast/ast.go
@@ -99,11 +99,14 @@ func (n Node) Valid() bool {
 
 var noNode = Node{}
 
+var Count = 0
+
 // Key returns the child nodes making the Key on a supported node. Panics
 // otherwise.
 // They are guaranteed to be all be of the Kind Key. A simple key would return
 // just one element.
 func (n *Node) Key() Iterator {
+	Count++
 	switch n.Kind {
 	case KeyValue:
 		value := n.Child()

```

Which shows:

```
old: 58 calls/op
new: 113 calls/op
```

May be onto something.

--- 

![image](https://user-images.githubusercontent.com/172804/120104670-9bfd0a80-c123-11eb-9c0d-e373072cb4a4.png)