@pelletier 2021-05-30
This seems to be the crux of the "performance regression" for `ReferenceFile/struct`: v2 has a bug (https://github.com/pelletier/go-toml/pull/547) that makes it skip over most of the document, so of course it would be much faster (but wrong).